[RFC] Introduce AccumulateGlobals
This prepares to replace compute_globals() with an object that always
reduces the same fixed-size buffer, and relies on the modules that
need to use the data to hold the logic about which steps need to fill
parts of that buffer so that they will be able to read valid global
values later.

Refs #2616

This used to run under I44a7193cc5492e3d6b37b896e16c6b438d2e28cc, but
for now we will give this a new Change-Id to avoid that this gets
rebased before we're sure it should.

Change-Id: Ic40a76faca43a5bb4b4b89d69823b6ea82134add