Make constraining elements use MicroState interfaces
The constraining elements were relying on the legacy t_state as input.
This commit changes this to accessing the parts of the MicroState
explicitly. Note that this required a change to do_constrain_first,
which causes an additional function call for the current do_md - since
it's only called once, I'm not worried about any performance issues.

Change-Id: Ibfdda35209bdf3cede4a9e86c6e8cd61d02dbcbe