MicroState is not based on t_state anymore
MicroState now owns the vectors for position, velocity and forces
directly. t_state* is only used in communication with the domain
decomposition element (and, temporarily, the shellfc element -
changing this element is not a priority, maybe we should turn
if off for now).

Looking at the current state, the question arises whether the
domain decomposition element should actually be the same entity
as the MicroState - to discuss further.

Change-Id: I9dd7b57f39026199f3d30465890d69004ba40280