Replaced dependence on Update object by xprime in finish_update
As we try to reorganize the data, it is helpful to avoid having
dependencies on large composite data structures, especially if only
a fraction of the data is actually used. This change replaces the
entire `Update` object as an input by only the x-prime vector, the
only part of the `Update` object actually used.

Change-Id: Ic36cc4bfa2487680d1b5f9bc54f0d176109a4d17