[RFC] Replace functionality of `Update` by `MicroState`
The `Update` class is a bit of a "catch-all" class for different
concepts related to the update phase. The concept of a catch-all
class is not fitting to an approach where data is distributed to
be owned by their respective algorithm objects. The current integrator
framework needs only one part of the `Update` class, the x-prime
vector. This change transfers this to the MicroState.

Ideally, the objects would keep copies of data they need privately
(e.g., the constrain algorithm keeps a reference position if needed).
The concept of the x-prime vector is, however, deeply incorporated in
the update algorithms, so it might be worth to keep this as a part of
the MicroState.

Change-Id: I5ac6026f9ef6333a7e4fe0c9ed6bbfd3146e44f0