[RFC] Scheduler approach
This change explores the possibility to express the integrator in a
scheduled way - i.e., the needed run methods are precomputed for
several steps, and then ran. Currently, this is arranged around the
neighbor searching steps - at each NS step (+ domdec, if needed),
all steps until the next NS step are precomputed.

Change-Id: Ifc8277e9df74add352f1470f889df0a608f3a158