Introduce Simulator base class
In view of having simulators differing based on user input rather
than a large single rigid simulator object that does everything,
this change introduces a Simulator interface. The current Simulator
class is renamed to LegacySimulator, and implements the interface.
Derived classes cannot be aggregate-initialized, so a constructor
has been added to LegacySimulator. This change paves to way to use
a builder for custom Simulator objects rather than the current
single implementation.

Change-Id: I6afa478e4d01f6d52b2ed7689acea51a0f69ad5c