Introduce do_simplemd() and SimpleSimulator
This introduces do_simplemd(), currently an identical copy of do_md(),
and SimpleSimulator, currently very similar to LegacySimulator.
In future commits, this will be stripped down to the bare minimum
to simplify the implementation of the new simulator framework.

A feature flag / environment variable `GMX_USE_MODULAR_SIMULATOR`
is introduced. If it is present, the SimulatorBuilder returns an
object of type SimpleSimulator instead of LegacySimulator, which
will in turn call do_simplemd() instead of do_md().

The validity of the do_md / simple_md equivalence is tested by the
tests introduced in I5a36ce0c, and can be checked by a simple diff
of src/gromacs/mdrun/{md,simple_md}.cpp.

Change-Id: I79031520238245ddbec97cba5b6642d57f809c84